+++ weed/util/log_buffer/sealed_buffer.go
